SBSA Updates for Bootwrapper support
- Deprecating Baremetal support for RDN1
- Removed RDN1 folders from pal_baremetal.
- Continuing baremetal support for FVP RDN2.
- Added Bootentry script to get control at BL33 and performs system init.
- Introduced TARGET_BM_BOOT macro to enable Bootwrapper
   support with current ACS infra.
- Included Memory map and MMU enable support.
- Aligned memory accesses for bootwrapper.
- Added PL011 UART driver code for Bootwrapper console prints.
- Moved RDN2 folder from pal_baremetal/FVP to pal_baremetal.
- Introduced a common folder for generic PAL APIs.
- Using val_print instead of pal_print inside val folder.
- Removed existing Makefiles.
- Cmake files for generation of .elf, .bin, .map , .dump files.
- Code improvements to ensure no compiler warnings and errors.
- Removal of unused variables.

+++ test_pool/mpam/operating_system/test_mpam006.c
@@ -37,7 +37,7 @@ static void payload(void)
     uint32_t msc_index;
     uint32_t rsrc_index;
     uint32_t llc_index;
-    uint32_t cache_identifier;
+    uint64_t cache_identifier;
     uint32_t cache_size;
     uint32_t max_pmg;
     uint32_t max_partid;
